chore(web): self-bundle Monaco, stop loading from jsdelivr CDN
@monaco-editor/react's default loader fetches Monaco core from
cdn.jsdelivr.net at first mount, which probed at ~5 s ttfb through
this user's Clash proxy and is incompatible with self-hosted /
air-gapped deploys (SPEC §13). Bundle Monaco from npm instead and
configure the React wrapper's loader to use the local copy.

- pnpm add monaco-editor.
- DiffView pulls in monaco-editor, sets MonacoEnvironment with
  Vite's ?worker import for editor.worker (skip language workers —
  read-only diff only needs Monarch tokenization), then
  loader.config({ monaco }).
- DiffSlice / payloadToDiff move out of DiffView into
  src/lib/payloadToDiff.ts so EventCard can import them
  synchronously without dragging Monaco into the eager bundle.
- EventCard React.lazy-loads DiffView and renders it inside
  <Suspense fallback="Loading diff editor…">. Users who never
  expand a diff never pay the Monaco cost.
- src/vite-env.d.ts adds the standard /// <reference types="vite/client" />
  so TypeScript recognises ?worker imports.

Build verification: production output splits cleanly:
- index chunk: 391 kB / 127 kB gzipped (no Monaco)
- DiffView chunk: 3.8 MB / 990 kB gzipped (Monaco), lazy-loaded
- ~30 per-language chunks loaded on demand by file extension

Closes #45
